HelpChannels: implement the !dormant command
Basically a wrapper around move_to_dormant which ensures the current
channel is in use. If it's not in-use, from the invoker's perspective,
the command silently fails (it does at least log). InChannelCheckFailure
was considered but it seemed like it'd be too spammy, especially if
there'd be a long list of allowed channels.