HelpChannels: fix unawaited coro warning for set_permissions
This was happening when attempting to schedule a task twice for a user.
Because the scheduler refuses to schedule a duplicate, the coroutine
is deallocated right away without being awaited (or closed explicitly
by `scheduled_task`).

To fix, any existing task is cancelled before scheduling. This also
means if somehow a user bypasses the lack of permissions, their
cooldown will be updated. However, it probably doesn't make a difference
as if they can bypass once, they likely can bypass again.